refactor: reorder methods
